kcse-telegram-bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pytest==7.4.0